+++ task/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.6
 
+- Allow sort by `modified`, display `modified` in task list view
 - Tap non-leaf project filters to fold/hide subproject filters
   + Non-leaf projects indicated by contributing task count in parentheses
   + Unfortunately, toggling a non-leaf project filter is less convenient